requirements.txt
selenium
unidecode
praw
python-dotenv
playsound
requests